requirements.txt
aiohttp==3.6.2
pyyaml==5.1.2
multidict==4.5.2
